docs: reclarify upgrade guide for deprecated label
As removing the labels is an operation extreamly complex we should warn
the users that the label was removed but still give them the option to
keep the label to upgrade from older versions to v1.7. During testing it
was found that following the existing upgrade guide could leave 2 Cilium
pods running per each node when performing a downgrade.

Signed-off-by: André Martins <andre@cilium.io>