policy: Give priotity for L7 redirect entries
Two selectors can select the same security identity, and the policy
can specify L7-rules for one of them, but not the other. While L7
enforcement would make this policy inconsistent, such a policy is
valid for L7 visibility. To ensure consistent behavior give priority
to redirect in the case the same map key has both a redirect and
non-redirect.

The implementation requires updates on a given security ID to happen
atomically. To achive this ConsumeMapChanges() is moved to the
endpoint policy, which can lock the selectorcache to guarantee that
concurrent updates on any new security ID have completed.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>